% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blackbird.R
\docType{data}
\name{blackbird}
\alias{blackbird}
\title{Blackbird combinator}
\format{
An object of class \code{help_files_with_topic} of length 1.
}
\usage{
blackbird
}
\description{
Blackbird combinator
}
\details{
#  x                  x
#  |                  |
#  v                  v
# c(x)               d(x)
#  |                  |
#  +->b(alpha,omega)<-+
#            |
#            v
#         a(alpha)
}
\keyword{datasets}
